ENH: Add vtkMRMLSliceLogicTest2 test
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=15289
git-svn-id: http://svn.slicer.org/Slicer4/trunk@15289 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLScalarVolumeDisplayNode.cxx
@@ -407,7 +407,7 @@ void vtkMRMLScalarVolumeDisplayNode::AddWindowLevelPreset(double window, double 
 //---------------------------------------------------------------------------
 int vtkMRMLScalarVolumeDisplayNode::GetNumberOfWindowLevelPresets()
 {
-  return this->WindowLevelPresets.size();
+  return static_cast<int>(this->WindowLevelPresets.size());
 }
 
 //---------------------------------------------------------------------------
